requirements.txt
requests==2.28.2
websockets~=10.4
configobj==5.0.8
colorlog~=6.7.0
sqlalchemy==2.0.2
lxml==4.9.2

Pillow~=9.4.0
qrcode==7.4.2
aiohttp~=3.8.3
setuptools==67.2.0
beautifulsoup4==4.11.2
overrides~=7.3.1
APScheduler==3.10.0
jieba~=0.42.1
packaging==23.0
nest_asyncio
rich~=13.3.1
tzlocal~=4.2
pydantic~=1.10.4
openai==0.26.5
aiofiles==23.1.0
PyYAML~=6.0
ConfigUpdater~=3.1.1
httpx==0.23.3
opencv-python~=4.7.0.68
numpy==1.24.2
pyzbar~=0.1.9
psutil~=5.9.4
table2ascii~=1.1.0